schemaexpr: move NameResolutionVisitor from sqlbase
Release note: None
