+++ pkg/sql/select_name_resolution.go
@@ -15,8 +15,8 @@ package sql
 
 import (
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/colinfo"
+	"github.com/cockroachdb/cockroach/pkg/sql/schemaexpr"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/sql/sqlbase"
 )
 
 // resolveNames walks the provided expression and resolves all names
@@ -27,5 +27,5 @@ func (p *planner) resolveNames(
 	if expr == nil {
 		return nil, nil
 	}
-	return sqlbase.ResolveNamesUsingVisitor(&p.nameResolutionVisitor, expr, source, ivarHelper, p.SessionData().SearchPath)
+	return schemaexpr.ResolveNamesUsingVisitor(&p.nameResolutionVisitor, expr, source, ivarHelper, p.SessionData().SearchPath)
 }
